fix(types): mypy 棘轮批四烧减——869→694（-175），零 suppression
质量卡 wt536-mypy4（基线 869，目标 ≤829）。全程未新增任何 type: ignore /
py.typed / mypy 配置改动；仅注解、返回类型、Optional 链与有依据的 cast。

烧减族清单：
1. BaseTool.execute 覆写（app/tools/ 15 文件，69 条）：补齐基类契约参数
   locale: str = "en"（executor 按 _accepts_kwarg 探测后传入，兼容旧签名）；
   随后暴露的 params 收窄 [override]（37 处）改为 params: BaseModel 入参 +
   函数首行 cast(具体 Schema, params)——依据：executor.py:870 以
   tool.parameters_schema(**arguments) 构造 validated_params 后才调用 execute。
2. redis await 联合（25 文件，61 条）：redis-py async 命令注解为
   Union[Awaitable[T], T]（sync/async 共享签名），新增
   app/core/redis_utils.ensure_awaitable(result: Awaitable[T] | T) -> Awaitable[T]
   运行期恒走 awaitable 分支（async 客户端返回值必为 awaitable），不改行为。
   连带修正暴露的下游：llm_quota evalsha 按 rate_limit.lua 返回 {int,int}
   cast list[int]；websocket 离线队列项为 json.dumps 字符串 cast str|None
   （websocket.py:516 推入侧佐证）；semantic_cache candidate_keys 分支联合注解。
   非 redis 三点：memory_use_selfcheck FastModelHook 别名改注
   Coroutine[Any, Any, str | None]（测试钩子均为 async）；db/error_handler
   retry_on_deadlock func 改 Callable[P, Awaitable[T]] 并去掉多余 cast。
3. max(d, key=d.get)（12 文件，19 条）：迭代源即字典本身，key 恒存在，
   改 key=lambda k: d[k]，行为等价、可静态求解。
4. dict-item / 注解失真（13 文件，22 条）：get_sector_mapping 值含
   list[str]、build_model_options 值含 bool、SURFACE_PAYLOAD_SCHEMAS 特判面
   值为 None、InferredEpisodicCandidate.evidence_refs 实载 float/bool/dict
   （下游 create_episodic_memory 为 Iterable[Any]）、state_aggregator fetcher
   注册表异构签名放宽 Callable[...]、_calculate_confidence_calibration 占位
   返回 str/bool、tool_summary 链路（validator→record_service）实含
   success_rate float、memory_jobs summary 混合值等——按运行期真实形状
   修正注解。
5. 可选依赖导入回退（8 文件，13 条 Cannot assign to a type）：mdx/prometheus/
   openai/jinja2/L1LightAurora/GroupFileService/RedisError 等 except 分支
   向导入类型名赋 None。改为模块级 Any 声明变量 + 别名导入回填，运行期
   语义不变，并顺带移除 4 处错码不匹配的旧 type: ignore[assignment]
   （其真实错误码为 [misc]，本就未被抑制）。

已知在途 FIX-253 四点位（aurora/runtime_v1/service.py 三处、
orchestration/orchestrator.py 一处）按卡规避未动。

跳过留裁决（2 条，行为修复超出本卡授权）：
- app/services/routing_parameter_proposal_service.py:197 await 了同步纯函数
  _compute_proposed_value（:315 def，返回 float|None）——运行期到达即
  TypeError，属潜伏 bug，需裁决改同步调用或改 async。
- app/orchestration/persistence_layer.py:313 proto metadata（map<string,str>）
  传入 preview.get("proposal_id")（Any|None）与 task_count int——缺键/类型
  不符时运行期 protobuf 报错，需裁决缺省值策略。

回归：tools/sse/cache/quota/websocket/signal/theater/health/spine/predictive/
weekly/policy/validator/plan_review/state_aggregator/memory lane/reflection
等 21 批约 1600 用例通过（bert_intent_classifier 36 个 setup error 为既有
环境问题——transformers 未装，主仓同现）。终跑 ratchet：errors: 694 / baseline: 694。

+++ backend/app/services/plan_execution_record_service.py
@@ -34,7 +34,7 @@ class PlanExecutionRecordService:
         validation_status: str,
         quality_score: float,
         criteria_results: dict[str, Any],
-        tool_summary: dict[str, int],
+        tool_summary: dict[str, float],
         issues: list[str],
         execution_intent_id: UUID | None = None,
     ) -> PlanExecutionRecord:
